Consolidating insert queries while maintaining uniqueness of the game creation and updating functions
