+++ erzahler/database/queries/game/insert-bridge-query.ts
