+++ erzahler/database/queries/game/insert-country-history-query.ts
