+++ erzahler/database/queries/game/insert-country-query.ts
