+++ erzahler/database/queries/game/insert-node-adjacency-query.ts
