+++ erzahler/database/queries/game/insert-node-query.ts
