+++ erzahler/database/queries/game/insert-province-history-query.ts
